official_bot/requirements.txt
# Telegram Bot Framework
python-telegram-bot==20.7

# HTTP Client
aiohttp==3.9.1
requests==2.31.0

# Database
sqlalchemy==2.0.23

# Password Hashing
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# Logging
colorlog==6.8.0

# Date/Time
python-dateutil==2.8.2

# Validation
validators==0.22.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
